Remove pprofile as req'd dep
